requirements.txt
openai
requests
feedparser
python-dotenv
python-telegram-bot
